refactor: extract language-specific logic and split enhanced_dependency_resolver
- Extract language-specific parts to languages/ directory
  - Create rust_enhanced_resolver.rs for Rust-specific logic
  - Create typescript_enhanced_resolver.rs for TypeScript-specific logic
- Split enhanced_dependency_resolver into smaller modules under dependency_resolver/
  - type_system.rs: Generic type system structures
  - constraint_solver.rs: Constraint solving engine
  - trait_database.rs: Trait definitions and implementations
  - associated_type_resolver.rs: Associated type resolution
  - lifetime_resolver.rs: Lifetime management
  - resolution_candidate.rs: Shadowing and resolution candidates
  - base_resolver.rs: DependencyResolver trait definition
- Update imports and references to use new modular structure
- Maintain all existing functionality and pass all tests

Resolves #110
