+++ crates/core/src/dependency_resolver/associated_type_resolver.rs
@@ -1,0 +1,63 @@
+use super::trait_database::{TraitDatabase, TraitImplementation};
+use super::type_system::TraitId;
+use crate::models::Type;
+use std::collections::HashMap;
+
+/// Associated Type Resolver
+#[derive(Debug, Clone)]
+pub struct AssociatedTypeResolver {
+    pub trait_database: TraitDatabase,
+    pub impl_database: ImplDatabase,
+}
+
+#[derive(Debug, Clone)]
+pub struct ImplDatabase {
+    pub implementations: HashMap<String, Vec<TraitImplementation>>,
+}
+
+impl Default for AssociatedTypeResolver {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl AssociatedTypeResolver {
+    pub fn new() -> Self {
+        Self {
+            trait_database: TraitDatabase::new(),
+            impl_database: ImplDatabase::new(),
+        }
+    }
+
+    pub fn resolve_associated_type(
+        &self,
+        trait_impl: &TraitImplementation,
+        assoc_name: &str,
+    ) -> Option<Type> {
+        trait_impl.associated_type_mappings.get(assoc_name).cloned()
+    }
+
+    pub fn project_type(
+        &self,
+        _base_type: &Type,
+        _trait_def: TraitId,
+        _assoc_name: &str,
+    ) -> Option<Type> {
+        // Implement type projection for associated types
+        None
+    }
+}
+
+impl Default for ImplDatabase {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl ImplDatabase {
+    pub fn new() -> Self {
+        Self {
+            implementations: HashMap::new(),
+        }
+    }
+}
