+++ crates/core/src/dependency_resolver/constraint_solver.rs
@@ -1,0 +1,74 @@
+use super::trait_database::TraitDatabase;
+use super::type_system::{Constraint, ConstraintError, TraitBound, TypeSubstitution, TypeVarId};
+use crate::models::Type;
+use std::collections::HashMap;
+
+/// Constraint solving engine for generic type constraints
+#[derive(Debug, Clone)]
+pub struct ConstraintSolver {
+    pub active_constraints: HashMap<TypeVarId, Vec<Constraint>>,
+    pub trait_database: TraitDatabase,
+}
+
+impl Default for ConstraintSolver {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl ConstraintSolver {
+    pub fn new() -> Self {
+        Self {
+            active_constraints: HashMap::new(),
+            trait_database: TraitDatabase::new(),
+        }
+    }
+
+    pub fn add_constraint(&mut self, type_var: TypeVarId, constraint: Constraint) {
+        self.active_constraints
+            .entry(type_var)
+            .or_default()
+            .push(constraint);
+    }
+
+    pub fn solve_constraints(&mut self) -> Result<TypeSubstitution, ConstraintError> {
+        // Basic constraint solving implementation
+        let substitution = TypeSubstitution {
+            type_vars: HashMap::new(),
+            lifetimes: HashMap::new(),
+        };
+
+        // Simplified solving logic - to be enhanced
+        for constraints in self.active_constraints.values() {
+            for constraint in constraints {
+                match constraint {
+                    Constraint::TraitBound { .. } => {
+                        // Check trait bounds
+                    }
+                    Constraint::Equality { .. } => {
+                        // Handle type equality
+                    }
+                    Constraint::Lifetime { .. } => {
+                        // Handle lifetime constraints
+                    }
+                    Constraint::Associated { .. } => {
+                        // Handle associated type constraints
+                    }
+                }
+            }
+        }
+
+        Ok(substitution)
+    }
+
+    pub fn check_trait_bounds(&self, type_args: &[Type], bounds: &[TraitBound]) -> bool {
+        // Basic trait bound checking - to be enhanced
+        for _bound in bounds {
+            // Check if each type argument satisfies the bound
+            for _type_arg in type_args {
+                // Implement trait bound validation
+            }
+        }
+        true
+    }
+}
