+++ crates/core/src/dependency_resolver/lifetime_resolver.rs
@@ -1,0 +1,46 @@
+use super::type_system::LifetimeScope;
+use crate::models::scope::ScopeId;
+use std::collections::HashMap;
+
+/// Lifetime Resolver for handling lifetime parameters
+#[derive(Debug, Clone)]
+pub struct LifetimeResolver {
+    pub lifetime_scopes: HashMap<ScopeId, LifetimeScope>,
+    pub borrow_checker: BorrowChecker,
+}
+
+#[derive(Debug, Clone)]
+pub struct BorrowChecker {
+    // Placeholder for borrow checking functionality
+}
+
+impl Default for LifetimeResolver {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl LifetimeResolver {
+    pub fn new() -> Self {
+        Self {
+            lifetime_scopes: HashMap::new(),
+            borrow_checker: BorrowChecker::new(),
+        }
+    }
+
+    pub fn add_lifetime_scope(&mut self, scope_id: ScopeId, scope: LifetimeScope) {
+        self.lifetime_scopes.insert(scope_id, scope);
+    }
+}
+
+impl Default for BorrowChecker {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl BorrowChecker {
+    pub fn new() -> Self {
+        Self {}
+    }
+}
