+++ crates/core/src/languages/rust/mod.rs
@@ -1,5 +1,6 @@
 pub mod formatter;
 pub mod rust_definition_collector;
 pub mod rust_dependency_resolver;
+pub mod rust_enhanced_resolver;
 pub mod rust_impl_collector;
 pub mod rust_usage_node_collector;
