+++ crates/core/src/languages/rust/rust_enhanced_resolver.rs
@@ -1,0 +1,200 @@
+use crate::dependency_resolver::{
+    AssociatedTypeResolver, ConstraintError, GenericTypeResolver,
+    LifetimeResolver, TypeParam,
+};
+use crate::enhanced_dependency_resolver::EnhancedDependencyResolver;
+use crate::method_resolver::MethodResolutionResult;
+use crate::models::{
+    scope::{ScopeId, SymbolTable},
+    Definition, Type, Usage,
+};
+use tree_sitter::Node;
+
+pub struct RustEnhancedResolver {
+    symbol_table: SymbolTable,
+    generic_type_resolver: GenericTypeResolver,
+    associated_type_resolver: AssociatedTypeResolver,
+    lifetime_resolver: LifetimeResolver,
+}
+
+impl RustEnhancedResolver {
+    pub fn new(symbol_table: SymbolTable) -> Self {
+        Self {
+            symbol_table,
+            generic_type_resolver: GenericTypeResolver::new(),
+            associated_type_resolver: AssociatedTypeResolver::new(),
+            lifetime_resolver: LifetimeResolver::new(),
+        }
+    }
+
+    /// Initialize method resolution for Rust code by analyzing impl blocks and traits
+    pub fn analyze_impl_blocks(
+        &mut self,
+        enhanced_resolver: &mut EnhancedDependencyResolver,
+        source_code: &str,
+        root_node: Node,
+    ) -> Result<(), String> {
+        let mut impl_collector =
+            crate::languages::rust::rust_impl_collector::RustImplCollector::new();
+
+        // Collect impl blocks
+        let impl_blocks = impl_collector.collect_impl_blocks(source_code, root_node)?;
+        for impl_block in impl_blocks {
+            enhanced_resolver
+                .method_resolver
+                .impl_block_analyzer
+                .add_impl_block(impl_block);
+        }
+
+        // Collect traits
+        let traits = impl_collector.collect_traits(source_code, root_node)?;
+        for trait_def in traits {
+            enhanced_resolver
+                .method_resolver
+                .trait_resolver
+                .add_trait(trait_def);
+        }
+
+        // Collect trait implementations
+        let trait_impls = impl_collector.collect_trait_impl_blocks(source_code, root_node)?;
+        for trait_impl in trait_impls {
+            enhanced_resolver
+                .method_resolver
+                .trait_resolver
+                .add_trait_impl(trait_impl);
+        }
+
+        Ok(())
+    }
+
+    /// Analyze generic type parameters in Rust code
+    pub fn analyze_generic_parameters(
+        &mut self,
+        source_code: &str,
+        root_node: Node,
+    ) -> Result<(), String> {
+        // Collect generic parameters from function definitions
+        let function_generics = self.collect_function_generics(source_code, root_node)?;
+        for (scope_id, params) in function_generics {
+            self.generic_type_resolver
+                .add_type_parameters(scope_id, params);
+        }
+
+        // Collect generic parameters from struct/enum definitions
+        let type_generics = self.collect_type_generics(source_code, root_node)?;
+        for (scope_id, params) in type_generics {
+            self.generic_type_resolver
+                .add_type_parameters(scope_id, params);
+        }
+
+        // Collect generic parameters from impl blocks
+        let impl_generics = self.collect_impl_generics(source_code, root_node)?;
+        for (scope_id, params) in impl_generics {
+            self.generic_type_resolver
+                .add_type_parameters(scope_id, params);
+        }
+
+        Ok(())
+    }
+
+    fn collect_function_generics(
+        &self,
+        _source_code: &str,
+        _root_node: Node,
+    ) -> Result<Vec<(ScopeId, Vec<TypeParam>)>, String> {
+        // Placeholder implementation
+        // TODO: Parse function signatures and extract generic parameters
+        Ok(Vec::new())
+    }
+
+    fn collect_type_generics(
+        &self,
+        _source_code: &str,
+        _root_node: Node,
+    ) -> Result<Vec<(ScopeId, Vec<TypeParam>)>, String> {
+        // Placeholder implementation
+        // TODO: Parse struct/enum definitions and extract generic parameters
+        Ok(Vec::new())
+    }
+
+    fn collect_impl_generics(
+        &self,
+        _source_code: &str,
+        _root_node: Node,
+    ) -> Result<Vec<(ScopeId, Vec<TypeParam>)>, String> {
+        // Placeholder implementation
+        // TODO: Parse impl blocks and extract generic parameters
+        Ok(Vec::new())
+    }
+
+    /// Resolve generic method calls with type parameter constraints
+    pub fn resolve_generic_method_call(
+        &self,
+        enhanced_resolver: &EnhancedDependencyResolver,
+        usage: &Usage,
+        source_code: &str,
+        root_node: Node,
+        definitions: &[Definition],
+    ) -> Option<MethodResolutionResult> {
+        // First try normal method resolution
+        if let Some(result) = enhanced_resolver
+            .method_resolver
+            .resolve_method_call(usage, source_code, root_node, definitions)
+        {
+            // Enhanced with generic constraint checking
+            if self.validate_generic_constraints(&result) {
+                return Some(result);
+            }
+        }
+
+        None
+    }
+
+    fn validate_generic_constraints(&self, _result: &MethodResolutionResult) -> bool {
+        // Placeholder for generic constraint validation
+        // TODO: Check that all type parameters satisfy their trait bounds
+        true
+    }
+
+    /// Resolve associated types in generic contexts
+    pub fn resolve_associated_type_usage(&self, usage: &Usage, _scope_id: ScopeId) -> Option<Type> {
+        // Check if this usage refers to an associated type
+        if usage.name.contains("::") {
+            let parts: Vec<&str> = usage.name.split("::").collect();
+            if parts.len() >= 2 {
+                let _trait_name = parts[0];
+                let _assoc_type_name = parts[1];
+                // TODO: Implement associated type resolution
+            }
+        }
+
+        None
+    }
+
+    /// Check for higher-ranked trait bounds (HRTB)
+    pub fn check_higher_ranked_bounds(
+        &self,
+        _usage: &Usage,
+        _scope_id: ScopeId,
+    ) -> Result<bool, ConstraintError> {
+        // Placeholder for HRTB checking
+        // TODO: Implement for<'a> Fn(&'a str) -> &'a str patterns
+        Ok(true)
+    }
+
+    pub fn get_generic_type_resolver(&self) -> &GenericTypeResolver {
+        &self.generic_type_resolver
+    }
+
+    pub fn get_generic_type_resolver_mut(&mut self) -> &mut GenericTypeResolver {
+        &mut self.generic_type_resolver
+    }
+
+    pub fn get_associated_type_resolver(&self) -> &AssociatedTypeResolver {
+        &self.associated_type_resolver
+    }
+
+    pub fn get_lifetime_resolver(&self) -> &LifetimeResolver {
+        &self.lifetime_resolver
+    }
+}
