+++ crates/core/src/languages/typescript/mod.rs
@@ -1,4 +1,5 @@
 pub mod formatter;
 pub mod typescript_definition_collector;
 pub mod typescript_dependency_resolver;
+pub mod typescript_enhanced_resolver;
 pub mod typescript_usage_node_collector;
