+++ crates/core/src/languages/typescript/typescript_enhanced_resolver.rs
@@ -1,0 +1,41 @@
+use crate::models::{
+    scope::{ScopeId, SymbolTable},
+    Definition, Type, Usage,
+};
+use tree_sitter::Node;
+
+pub struct TypeScriptEnhancedResolver {
+    symbol_table: SymbolTable,
+}
+
+impl TypeScriptEnhancedResolver {
+    pub fn new(symbol_table: SymbolTable) -> Self {
+        Self { symbol_table }
+    }
+
+    /// Analyze TypeScript-specific type parameters and generics
+    pub fn analyze_type_parameters(
+        &mut self,
+        _source_code: &str,
+        _root_node: Node,
+    ) -> Result<(), String> {
+        // TODO: Implement TypeScript generic analysis
+        Ok(())
+    }
+
+    /// Resolve TypeScript interface types and inheritance
+    pub fn resolve_interface_type(&self, _usage: &Usage, _scope_id: ScopeId) -> Option<Type> {
+        // TODO: Implement TypeScript interface resolution
+        None
+    }
+
+    /// Handle TypeScript module resolution and imports
+    pub fn resolve_module_import(
+        &self,
+        _usage: &Usage,
+        _definitions: &[Definition],
+    ) -> Option<Definition> {
+        // TODO: Implement TypeScript module import resolution
+        None
+    }
+}
